[Test] Test rosbag data on Autoware ros2 node
In order to write the launch file for the multi-modal perception stack
in a manner that conforms with test-driven development (TDD), we must first
ensure that the rosbags that we are using for the TDD approach contain
LiDAR data that is accessible to autoware ros2 nodes. These tests are
used to confirm that.

Partially resolves: Issue #39 in tartan_carpet project titled
Multi-Modal Perception Stack Autoware Implementation